requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# PDF Processing
PyPDF2==3.0.1

# Machine Learning & Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Search & Ranking
rank-bm25==0.2.2

# Database & Storage
sqlite-utils==3.35.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Development
black==23.11.0